Replace --network host with -p 8123:8123 for Home Assistant Docker installation (#5678)
This commit updates the Home Assistant Docker installation instructions to use the -p flag instead of --network host. By using -p 8123:8123, the container's services will be exposed only on the specified port and IP address, limiting the potential attack surface of the container. This change improves the security and isolation of the Home Assistant Docker installation.